# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type PropertyStatus = typeof PropertyStatus[keyof typeof PropertyStatus];


export const PropertyStatus = {
  available: 'available',
  pending: 'pending',
  rented: 'rented',
  paused: 'paused',
  hidden: 'hidden',
} as const;

export interface OwnerSummary {
  name: string;
  phone: string;
  initials: string;
  verified: boolean;
}

export interface Property {
  id: string;
  title: string;
  type: string;
  location: string;
  district: string;
  neighborhood: string;
  rent: number;
  advanceMonths: number;
  bedrooms: number;
  bathrooms: number;
  /** @nullable */
  size?: number | null;
  image: string;
  images?: string[];
  description?: string;
  amenities: string[];
  status: PropertyStatus;
  verified: boolean;
  owner: OwnerSummary;
  views: number;
  leads: number;
  createdAt: string;
  /** @nullable */
  lastVerifiedAt?: string | null;
}

export interface PropertyInput {
  /** @minLength 1 */
  title: string;
  type: string;
  district: string;
  neighborhood: string;
  address?: string;
  /** @minimum 0 */
  rent: number;
  /** @minimum 1 */
  advanceMonths: number;
  /** @minimum 0 */
  bedrooms: number;
  /** @minimum 0 */
  bathrooms: number;
  description: string;
  amenities: string[];
  image?: string;
}

export type PropertyUpdateStatus = typeof PropertyUpdateStatus[keyof typeof PropertyUpdateStatus];


export const PropertyUpdateStatus = {
  available: 'available',
  pending: 'pending',
  rented: 'rented',
  paused: 'paused',
  hidden: 'hidden',
} as const;

export interface PropertyUpdate {
  title?: string;
  /** @minimum 0 */
  rent?: number;
  status?: PropertyUpdateStatus;
  description?: string;
  amenities?: string[];
}

export interface TenantDashboard {
  savedProperties: number;
  savedSearches: number;
  recentViews: number;
  messages: number;
  recommended: Property[];
  nearby: Property[];
  recentlyViewed: Property[];
}

export type OwnerDashboardWeeklyViewsItem = {
  label: string;
  value: number;
};

export interface Activity {
  id: string;
  text: string;
  time: string;
  kind: string;
}

export interface OwnerDashboard {
  totalProperties: number;
  activeListings: number;
  totalViews: number;
  leads: number;
  weeklyViews: OwnerDashboardWeeklyViewsItem[];
  properties: Property[];
  activity: Activity[];
}

export type VerificationItemStatus = typeof VerificationItemStatus[keyof typeof VerificationItemStatus];


export const VerificationItemStatus = {
  pending: 'pending',
  under_review: 'under_review',
  verified: 'verified',
  rejected: 'rejected',
} as const;

export interface VerificationItem {
  id: string;
  owner: string;
  phone: string;
  submittedAt: string;
  status: VerificationItemStatus;
}

export interface AdminDashboard {
  totalUsers: number;
  tenants: number;
  owners: number;
  activeProperties: number;
  pendingVerification: number;
  monthlyRevenue: number;
  reportedListings: number;
  verificationQueue: VerificationItem[];
  moderationQueue: Property[];
}

export interface SavedSearch {
  id: string;
  name: string;
  location: string;
  summary: string;
  matches: number;
  updatedAt: string;
}

export interface SavedSearchInput {
  name: string;
  location: string;
  summary: string;
}

export type LeadStatus = typeof LeadStatus[keyof typeof LeadStatus];


export const LeadStatus = {
  new: 'new',
  contacted: 'contacted',
  viewing: 'viewing',
  closed: 'closed',
} as const;

export interface Lead {
  id: string;
  name: string;
  property: string;
  message: string;
  source: string;
  status: LeadStatus;
  createdAt: string;
}

export interface PropertyReport {
  id: string;
  propertyId: string;
  reason: string;
  createdAt: string;
}

export interface PropertyReportInput {
  propertyId: string;
  reason: string;
}

export interface ContactMessageInput {
  /** @minLength 2 */
  name: string;
  /** @minLength 3 */
  email: string;
  /** @minLength 1 */
  topic: string;
  /** @minLength 5 */
  message: string;
}

export interface ContactMessage {
  id: string;
  name: string;
  email: string;
  topic: string;
  message: string;
  createdAt: string;
}

export type ListPropertiesParams = {
location?: string;
type?: string;
maxRent?: number;
minRent?: number;
bedrooms?: number;
advanceMonths?: number;
amenity?: string;
sort?: ListPropertiesSort;
};

export type ListPropertiesSort = typeof ListPropertiesSort[keyof typeof ListPropertiesSort];


export const ListPropertiesSort = {
  recommended: 'recommended',
  lowest: 'lowest',
  highest: 'highest',
  newest: 'newest',
  viewed: 'viewed',
} as const;