# lib/api-zod/src/generated/types/contactMessageInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ContactMessageInput {
  /** @minLength 2 */
  name: string;
  /** @minLength 3 */
  email: string;
  /** @minLength 1 */
  topic: string;
  /** @minLength 5 */
  message: string;
}